Nuclear Attraction Code is working for current H2 test

+++ include/utils.h
@@ -20,4 +20,4 @@ double min(double, double);
 
 double max(double, double);
 
-double boys(int, int);
+double boys(double, double);
